+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 7.1.1 2024-11-28
+
+* Fixes and improvements for metadata management form
+
+## 7.1.0 2024-11-26
+
+* Added user authentification
+* Added new area for metadata management
+
 ## 7.0.0 2024-09-13
 
 * Initial commit: The template-engine replaces the old liferay based frontend ([Open CoDE](https://gitlab.opencode.de/fitko/govdata/GovDataPortal), [GitHub](https://github.com/GovDataOfficial/GovDataPortal))